build: Update compiled jars' filenames to include Git version info
The `v` in the Git tags gets stripped from the filename to keep it consistent with previous releases.
This depends on releasing prereleases on Git though, with at least a tag, though preferably just a release.
We need to add the MC version to the Git tag as well, and this requires separate releases per MC version to have the correct MC version in the filename.