[Feature]: Overhauled Extended and Gems module (#954)
This adds a deprecated module. This module is currently still activated
instantly so people dont actually lose their progress. It can be
deactivitated using the following pack names in the settings file:
 "forgero:extended-handles-deprecated",
 "forgero:extended-swords-deprecated",
 "forgero:extended-tools-deprecated",
 "forgero:material-deprecated",
 "forgero:trinkets-deprecated"

With this module certain things have been moved to deprecated and we do
not suggest using this:
- refined/mastercrafted handle variants
- polar bear / enderdragon scale 
- alot of old extended stuff
- old gems have been remade

Couple of things to note are: 
Some extended stuff is still broken but merging should be fine if we fix
(throwing, reach)
Gems should be fully good to go 
Do we want to activate this when starting or actually deactivate?

---------

Co-authored-by: sigmund <sigmundsgranaas@gmail.com>
Co-authored-by: Sigmund Granaas Sandring <46557903+SigmundGranaas@users.noreply.github.com>