[Patch]: Weapon types (#1117)
Implements Swordlike, spear and Blunt weapon as different weapon types.
* War hammer is the default for blunt weapons
* Spear is the default for spear
* Knife is the default for sword_like

We could add clubs into the blunt weapon. Tridents will probably also be
grouped into the spears type. The type system with defaulted weapons
needs to be reworked. It is clearly pushed to the limits now, and the
process for selecting which items should be the defaults is just
horrible to work with.

Check out the weapons by writing: `/forgero extended_hall`