THIS IS AI GENERATED CODE (IT WORKS THO)
config isnt set correctly, changed the code up to work just like vanilla with each block mined adding to the exhaustionlevel that then later depletes hunger.

Tested this with appleskin mod to show exhaustion levels -> 50 weight is base (0.005 just like vanilla), chosen because of the median of all our base pickaxes is 47 so chose the 50 as a nice middlepoint.